requirements.txt
Django==3.1.4
django-environ==0.4.5
ipython==7.19.0
Pillow==8.0.1
tweepy==3.10.0
